docs: combine appended title sections
- rename storybook story
