+++ packages/eui/src-docs/src/views/color_picker/color_palette_picker.js
@@ -1,4 +1,6 @@
 import React, { useState } from 'react';
+import { css } from '@emotion/react';
+
 import {
   euiPaletteColorBlind,
   euiPaletteForStatus,
@@ -10,32 +12,50 @@ import {
   EuiSpacer,
   EuiCode,
   EuiColorPalettePicker,
+  EuiText,
+  EuiFlexGroup,
 } from '../../../../src/components/';
 
 import { DisplayToggles } from '../form_controls/display_toggles';
 
-const palettes = [
+/** Text wrapper to remove text-decoration on appended text */
+const AppendedTitleText = ({ label }) => (
+  <EuiText color="subdued" size="xs">
+    <span
+      css={css`
+        display: inline-block;
+      `}
+    >
+      {label}
+    </span>
+  </EuiText>
+);
+
+const getPalettes = (appendTitles) => [
   {
     value: 'pallette_1',
-    title: 'EUI color blind (fixed)',
+    title: 'EUI color blind',
     palette: euiPaletteColorBlind(),
+    append: appendTitles && <AppendedTitleText label="fixed" />,
     type: 'fixed',
   },
   {
     value: 'pallette_2',
-    title: 'EUI palette for status (gradient)',
+    title: 'EUI palette for status',
     palette: euiPaletteForStatus(5),
+    append: appendTitles && <AppendedTitleText label="gradient" />,
     type: 'gradient',
   },
   {
     value: 'pallette_3',
-    title: 'EUI palette for temperature (fixed)',
+    title: 'EUI palette for temperature',
     palette: euiPaletteForTemperature(5),
+    append: appendTitles && <AppendedTitleText label="fixed" />,
     type: 'fixed',
   },
   {
     value: 'pallette_4',
-    title: 'Grayscale (gradient with stops)',
+    title: 'Grayscale',
     palette: [
       {
         stop: 100,
@@ -54,11 +74,12 @@ const palettes = [
         color: 'black',
       },
     ],
+    append: appendTitles && <AppendedTitleText label="gradient with stops" />,
     type: 'gradient',
   },
   {
     value: 'pallette_5',
-    title: 'Grayscale (fixed with stops)',
+    title: 'Grayscale',
     palette: [
       {
         stop: 100,
@@ -77,6 +98,7 @@ const palettes = [
         color: 'black',
       },
     ],
+    append: appendTitles && <AppendedTitleText label="fixed with stops" />,
     type: 'fixed',
   },
   {
@@ -88,23 +110,35 @@ const palettes = [
 
 export default () => {
   const [selectionDisplay, setSelectionDisplay] = useState(false);
+  const [showAppendedTitles, setShowAppendedTitles] = useState(false);
   const [pallette, setPallette] = useState('pallette_1');
 
   return (
     <>
-      <EuiSwitch
-        label={
-          <span>
-            Display selected item as a <EuiCode>title</EuiCode>
-          </span>
-        }
-        checked={selectionDisplay}
-        onChange={() => setSelectionDisplay(!selectionDisplay)}
-      />
+      <EuiFlexGroup>
+        <EuiSwitch
+          label={
+            <span>
+              Display selected item as a <EuiCode>title</EuiCode>
+            </span>
+          }
+          checked={selectionDisplay}
+          onChange={() => setSelectionDisplay(!selectionDisplay)}
+        />
+        <EuiSwitch
+          label={
+            <span>
+              Display <EuiCode>append</EuiCode> element on title
+            </span>
+          }
+          checked={showAppendedTitles}
+          onChange={() => setShowAppendedTitles(!showAppendedTitles)}
+        />
+      </EuiFlexGroup>
       <EuiSpacer />
       <DisplayToggles canPrepend={true} canAppend={true} canReadOnly={false}>
         <EuiColorPalettePicker
-          palettes={palettes}
+          palettes={getPalettes(showAppendedTitles)}
           onChange={setPallette}
           valueOfSelected={pallette}
           selectionDisplay={selectionDisplay ? 'title' : 'palette'}
