+++ packages/website/docs/components/forms/color_selection.mdx
@@ -35,126 +35,10 @@ Use **EuiColorPalettePicker** to select palettes to apply colors to data visuali
 
 Use the `palettes` prop to pass your palettes as an array `strings` or an array of `ColorStops` in the form of `{ stop: number, color: string }`. For each object, you should pass a palette (array of hex values) and specify the `type`. Use `fixed` palettes for categorical data and `gradient` palettes for continuous data.
 
-```tsx interactive
-import React, { useState } from 'react';
-import {
-  euiPaletteColorBlind,
-  euiPaletteForStatus,
-  euiPaletteForTemperature,
-  EuiSwitch,
-  EuiSpacer,
-  EuiCode,
-  EuiColorPalettePicker,
-} from '@elastic/eui';
-
-const palettes = [
-  {
-    value: 'pallette_1',
-    title: 'EUI color blind (fixed)',
-    palette: euiPaletteColorBlind(),
-    type: 'fixed',
-  },
-  {
-    value: 'pallette_2',
-    title: 'EUI palette for status (gradient)',
-    palette: euiPaletteForStatus(5),
-    type: 'gradient',
-  },
-  {
-    value: 'pallette_3',
-    title: 'EUI palette for temperature (fixed)',
-    palette: euiPaletteForTemperature(5),
-    type: 'fixed',
-  },
-  {
-    value: 'pallette_4',
-    title: 'Grayscale (gradient with stops)',
-    palette: [
-      {
-        stop: 100,
-        color: 'white',
-      },
-      {
-        stop: 250,
-        color: 'lightgray',
-      },
-      {
-        stop: 320,
-        color: 'gray',
-      },
-      {
-        stop: 470,
-        color: 'black',
-      },
-    ],
-    type: 'gradient',
-  },
-  {
-    value: 'pallette_5',
-    title: 'Grayscale (fixed with stops)',
-    palette: [
-      {
-        stop: 100,
-        color: 'white',
-      },
-      {
-        stop: 250,
-        color: 'lightgray',
-      },
-      {
-        stop: 320,
-        color: 'gray',
-      },
-      {
-        stop: 470,
-        color: 'black',
-      },
-    ],
-    type: 'fixed',
-  },
-  {
-    value: 'custom',
-    title: 'Plain text as a custom option',
-    type: 'text',
-  },
-];
-
-export default () => {
-  const [selectionDisplay, setSelectionDisplay] = useState(false);
-  const [pallette, setPallette] = useState('pallette_1');
-
-  return (
-    <>
-      <EuiSwitch
-        label={
-          <span>
-            Display selected item as a <EuiCode>title</EuiCode>
-          </span>
-        }
-        checked={selectionDisplay}
-        onChange={() => setSelectionDisplay(!selectionDisplay)}
-      />
-      <EuiSpacer />
-      <DisplayToggles canPrepend={true} canAppend={true} canReadOnly={false}>
-        <EuiColorPalettePicker
-          palettes={palettes}
-          onChange={setPallette}
-          valueOfSelected={pallette}
-          selectionDisplay={selectionDisplay ? 'title' : 'palette'}
-        />
-      </DisplayToggles>
-    </>
-  );
-};
-```
-
-## Color palette picker - Appended titles
-
 Each of the `palettes`, excluding `type='text'` palettes, can use the `append` prop to append an element to the right of the title.
 
 ```tsx interactive
 import React, { useState } from 'react';
-import { css } from '@emotion/react';
 import {
   euiPaletteColorBlind,
   euiPaletteForStatus,
@@ -162,22 +46,9 @@ import {
   EuiSwitch,
   EuiSpacer,
   EuiCode,
-  EuiText,
   EuiColorPalettePicker,
+  EuiFlexGroup
 } from '@elastic/eui';
-
-/** Text wrapper to remove text-decoration on appended text */
-const AppendedTitleText = ({ label }) => (
-  <EuiText color="subdued" size="xs">
-    <span
-      css={css`
-        display: inline-block;
-      `}
-    >
-      {label}
-    </span>
-  </EuiText>
-);
 
 const getPalettes = (appendTitles) => [
   {
@@ -256,27 +127,53 @@ const getPalettes = (appendTitles) => [
   },
 ];
 
+/** Text wrapper to remove text-decoration on appended text */
+const AppendedTitleText = ({ label }) => (
+  <EuiText color="subdued" size="xs">
+    <span
+      css={css`
+        display: inline-block;
+      `}
+    >
+      {label}
+    </span>
+  </EuiText>
+);
+
 export default () => {
-  const [showAppendedTitles, setShowAppendedTitles] = useState(true);
+  const [selectionDisplay, setSelectionDisplay] = useState(false);
+  const [showAppendedTitles, setShowAppendedTitles] = useState(false);
   const [pallette, setPallette] = useState('pallette_1');
 
   return (
     <>
-      <EuiSwitch
-        label={
-          <span>
-            Display <EuiCode>append</EuiCode> element on title
-          </span>
-        }
-        checked={showAppendedTitles}
-        onChange={() => setShowAppendedTitles(!showAppendedTitles)}
-      />
+      <EuiFlexGroup>
+        <EuiSwitch
+          label={
+            <span>
+              Display selected item as a <EuiCode>title</EuiCode>
+            </span>
+          }
+          checked={selectionDisplay}
+          onChange={() => setSelectionDisplay(!selectionDisplay)}
+        />
+        <EuiSwitch
+          label={
+            <span>
+              Display <EuiCode>append</EuiCode> element on title
+            </span>
+          }
+          checked={showAppendedTitles}
+          onChange={() => setShowAppendedTitles(!showAppendedTitles)}
+        />
+      </EuiFlexGroup>
       <EuiSpacer />
       <DisplayToggles canPrepend={true} canAppend={true} canReadOnly={false}>
         <EuiColorPalettePicker
           palettes={getPalettes(showAppendedTitles)}
           onChange={setPallette}
           valueOfSelected={pallette}
+          selectionDisplay={selectionDisplay ? 'title' : 'palette'}
         />
       </DisplayToggles>
     </>
